Effects

Description

Tipo de datos Effects

data Effects

Constructors

Say String

Up

Down

Left

Right

Backwards

UpsideDown

Negative

Delay Integer

Color Color

Repeat Integer [Effects]

Forever [Effects]

Funciones que operan sobre el tipo Effects

stringToPixel :: String -> Map Char Pixels -> Pixels

Función que convierte una palabra a un Pixel de color White.

changeColor :: Pixels -> Color -> Pixels

Función que cambia el color de un Pixels.

up :: Pixels -> Pixels

Función que desplaza una hilera hacia arriba.

 up (font 'A')
*   *
*   *
*   *
*****
*   *
*   *
 *** 

down :: Pixels -> Pixels

Función que desplaza una hilera hacia abajo.

left :: Pixels -> Pixels

Función que desplaza una hilera hacia la izquierda.

 left (font 'A')
***  
   **
   **
   **
*****
   **
   **

right :: Pixels -> Pixels

Función que desplaza una hilera hacia la derecha.

upsideDown :: Pixels -> Pixels

Función que invierte el orden de la filas.

 upsideDown (font 'A')
*   *
*   *
*****
*   *
*   *
*   *
 *** 

backwards :: Pixels -> Pixels

Función que invierte el orden de las columnas.

 backwards (font 'B')
 ****
*   *
*   *
 ****   
*   *
*   *
 ****

negative :: Pixels -> Pixels

Función que intercambia espacios en blanco por asteriscos y viceversa.

 negative (font 'A')
*   *
 *** 
 *** 
 *** 
     
 *** 
 ***