Pixels

Description

Definición del tipo Pixels y Pixel.

El tipo Pixels.

data Pixels

Constructors

Pixels

Fields

color :: Color

dots :: [[Pixel]]

data Pixel

Constructors

Pixel

Fields

on :: Boolean

Valor neutro del tipo Pixels.

defaultP :: Pixels

Valor neutro del tipo Pixels

Funciones que operan sobre los Pixels.

font :: Map Char Pixels -> Char -> Pixels

Función que encuentra el pixel correspondiente a un caracter.

readFont :: Handle -> IO (Map Char Pixels)

Función que lee cada letra y su respectiva representació en Pixels